Move ipv6loopback to helper package and make public
